fs, kernel: checkout to msm-3.10

+++ kernel/time.c
@@ -705,7 +705,7 @@ struct timespec timespec_add_safe(const struct timespec lhs,
 	struct timespec res;
 
 	set_normalized_timespec(&res, lhs.tv_sec + rhs.tv_sec,
-				(s64)lhs.tv_nsec + rhs.tv_nsec);
+				lhs.tv_nsec + rhs.tv_nsec);
 
 	if (res.tv_sec < lhs.tv_sec || res.tv_sec < rhs.tv_sec)
 		res.tv_sec = TIME_T_MAX;
